requirements.txt
tqdm
torch
torchvision
nni
thop
pillow
numpy